# Optimize `l=~".+"` matcher