Since dot is matching newline now, `l=~".+"` is "any non empty label value", and #14144 added a specific method in the index for that so we don't need to run the matcher on each one of the label values.

I only ran the relevant benchmarks:
```
goos: darwin
goarch: arm64
pkg: github.com/prometheus/prometheus/tsdb
cpu: Apple M3 Pro
                                                                     │   dp_main    │             dp_shortcut             │
                                                                     │    sec/op    │    sec/op     vs base               │
Querier/Head/PostingsForMatchers/i=~".+"-12                            9.203m ± 15%   4.989m ± 16%  -45.78% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-12                   9.130m ± 69%   4.813m ±  7%  -47.29% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-12              8.374m ±  9%   4.925m ±  8%  -41.18% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-12       8.552m ± 21%   5.141m ± 10%  -39.89% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-12     13.31m ± 15%   10.04m ±  9%  -24.56% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-12   18.65m ±  3%   15.62m ±  4%  -16.26% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-12   298.9n ±  3%   303.1n ± 11%        ~ (p=0.240 n=6)
geomean                                                                2.391m         1.617m        -32.37%

                                                                     │   dp_main    │              dp_shortcut              │
                                                                     │     B/op     │     B/op      vs base                 │
Querier/Head/PostingsForMatchers/i=~".+"-12                            12.23Mi ± 0%   10.70Mi ± 0%  -12.52% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-12                   12.23Mi ± 0%   10.70Mi ± 0%  -12.52% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-12              12.23Mi ± 0%   10.70Mi ± 0%  -12.52% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-12       12.23Mi ± 0%   10.70Mi ± 0%  -12.52% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-12     14.96Mi ± 0%   13.43Mi ± 0%  -10.24% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-12   18.14Mi ± 0%   16.61Mi ± 0%   -8.44% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-12     48.00 ± 0%     48.00 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                                2.234Mi        2.012Mi        -9.92%
¹ all samples are equal

                                                                     │  dp_main   │             dp_shortcut             │
                                                                     │ allocs/op  │ allocs/op   vs base                 │
Querier/Head/PostingsForMatchers/i=~".+"-12                            8.000 ± 0%   6.000 ± 0%  -25.00% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/i=~".+",j=~"X.+"-12                   12.00 ± 0%   10.00 ± 0%  -16.67% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",j="foo"-12              15.00 ± 0%   13.00 ± 0%  -13.33% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!="2",j="foo"-12       20.00 ± 0%   18.00 ± 0%  -10.00% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~"2.*",j="foo"-12     92.00 ± 0%   90.00 ± 0%   -2.17% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="1",i=~".+",i!~".*2.*",j="foo"-12   95.00 ± 0%   93.00 ± 0%   -2.11% (p=0.002 n=6)
Querier/Head/PostingsForMatchers/n="X",i=~".+",i!~".*2.*",j="foo"-12   3.000 ± 0%   3.000 ± 0%        ~ (p=1.000 n=6) ¹
geomean                                                                18.55        16.64       -10.31%
¹ all samples are equal
```


<!--
    Please give your PR a title in the form "area: short description".  For example "tsdb: reduce disk usage by 95%"

    If your PR is to fix an issue, put "Fixes #issue-number" in the description.

    Don't forget!

    - Please sign CNCF's Developer Certificate of Origin and sign-off your commits by adding the -s / --signoff flag to `git commit`. See https://github.com/apps/dco for more information.

    - If the PR adds or changes a behaviour or fixes a bug of an exported API it would need a unit/e2e test.

    - Where possible use only exported APIs for tests to simplify the review and make it as close as possible to an actual library usage.

    - Performance improvements would need a benchmark test to prove it.

    - All exposed objects should have a comment.

    - All comments should start with a capital letter and end with a full stop.
 -->